refactor(srm): 优化采购请求服务实现中的代码结构
- 简化了 SrmPurchaseRequestServiceImpl 中的代码结构
- 使用 lambda 表达式优化了 toBean 方法的调用，使代码更加简洁

+++ yudao-module-srm/yudao-module-srm-biz/src/main/java/cn/iocoder/yudao/module/srm/controller/admin/purchase/vo/order/req/SrmPurchaseOrderSaveJsonReqVO.java
@@ -22,7 +22,7 @@ public class SrmPurchaseOrderSaveJsonReqVO {
     private List<@Valid Item> items;
 
     @Schema(description = "版本号")
-    private Long version;
+    private Integer version;
 
     @Data
     public static class Item {
@@ -44,7 +44,7 @@ public class SrmPurchaseOrderSaveJsonReqVO {
         private Integer totalCompletionPassCount;
 
         @Schema(description = "版本号")
-        private Long version;
+        private Integer version;
     }
 
 }
